I can fix the `NSInternalInconsistencyException` by creating a `main.swift` file.
