+++ bitchat/main.swift
@@ -1,0 +1,36 @@
+//
+// main.swift
+// bitchat
+//
+// This is free and unencumbered software released into the public domain.
+// For more information, see <https://unlicense.org>
+//
+
+import SwiftUI
+import UserNotifications
+
+@main
+struct BitchatApp: App {
+    @StateObject private var chatViewModel = ChatViewModel()
+    #if os(iOS)
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+    #endif
+
+    init() {
+        UNUserNotificationCenter.current().delegate = NotificationDelegate.shared
+    }
+
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+                .environmentObject(chatViewModel)
+                .onAppear {
+                    NotificationDelegate.shared.chatViewModel = chatViewModel
+                }
+        }
+        #if os(macOS)
+        .windowStyle(.hiddenTitleBar)
+        .windowResizability(.contentSize)
+        #endif
+    }
+}
